[Web] split jslib
